--- modin/data_management/partitioning/axis_partition/__init__.py
@@ -1,25 +1,0 @@
-from .pandas_on_ray import (
-    PandasOnRayAxisPartition,
-    PandasOnRayColumnPartition,
-    PandasOnRayRowPartition,
-)
-from .pandas_on_python import (
-    PandasOnPythonAxisPartition,
-    PandasOnPythonColumnPartition,
-    PandasOnPythonRowPartition,
-)
-from .dask import DaskAxisPartition, DaskColumnPartition, DaskRowPartition
-from .utils import split_result_of_axis_func_pandas
-
-__all__ = [
-    "PandasOnRayAxisPartition",
-    "PandasOnRayColumnPartition",
-    "PandasOnRayRowPartition",
-    "PandasOnPythonAxisPartition",
-    "PandasOnPythonColumnPartition",
-    "PandasOnPythonRowPartition",
-    "DaskAxisPartition",
-    "DaskColumnPartition",
-    "DaskRowPartition",
-    "split_result_of_axis_func_pandas",
-]
