--- modin/data_management/partitioning/partition_collections/__init__.py
@@ -1,11 +1,0 @@
-from .base_block_partitions import BaseBlockPartitions
-from .ray_block_partitions import RayBlockPartitions
-from .python_block_partitions import PythonBlockPartitions
-from .dask import DaskBlockPartitions
-
-__all__ = [
-    "BaseBlockPartitions",
-    "RayBlockPartitions",
-    "PythonBlockPartitions",
-    "DaskBlockPartitions",
-]
