+++ modin/engines/dask/pandas_on_dask_delayed/axis_partition.py
@@ -1,9 +1,9 @@
 import pandas
 import dask
 
-from .axis_partition import BaseAxisPartition
-from ...partitioning.remote_partition import DaskRemotePartition
-from .utils import split_result_of_axis_func_pandas
+from modin.engines.base.axis_partition import BaseAxisPartition
+from modin.data_management.utils import split_result_of_axis_func_pandas
+from .remote_partition import DaskRemotePartition
 
 
 class DaskAxisPartition(BaseAxisPartition):
