+++ modin/engines/ray/pandas_on_ray/remote_partition.py
@@ -5,8 +5,8 @@ from __future__ import print_function
 import pandas
 import ray
 
-from .base_remote_partition import BaseRemotePartition
-from .utils import length_fn_pandas, width_fn_pandas
+from modin.engines.base.remote_partition import BaseRemotePartition
+from modin.data_management.utils import length_fn_pandas, width_fn_pandas
 
 
 class PandasOnRayRemotePartition(BaseRemotePartition):
