+++ modin/pandas/indexing.py
@@ -144,8 +144,8 @@ class _LocationIndexerBase(object):
     """
 
     def __init__(self, ray_df: DataFrame):
-        self.dm = ray_df._query_compiler
-        self.is_view = hasattr(self.dm, "is_view")
+        self.qc = ray_df._query_compiler
+        self.is_view = hasattr(self.qc, "is_view")
 
         self.row_scaler = False
         self.col_scaler = False
@@ -154,17 +154,17 @@ class _LocationIndexerBase(object):
         self, row_lookup: pandas.Index, col_lookup: pandas.Index, ndim: int
     ):
         if self.is_view:
-            dm_view = self.dm.__constructor__(self.dm.data, row_lookup, col_lookup)
+            qc_view = self.qc.__constructor__(self.qc.data, row_lookup, col_lookup)
         else:
-            dm_view = self.dm.view(row_lookup, col_lookup)
+            qc_view = self.qc.view(row_lookup, col_lookup)
 
         if ndim == 2:
-            return DataFrame(query_compiler=dm_view)
+            return DataFrame(query_compiler=qc_view)
         elif ndim == 0:
-            return dm_view.squeeze(ndim=0)
+            return qc_view.squeeze(ndim=0)
         else:
             single_axis = 1 if self.col_scaler else 0
-            return dm_view.squeeze(ndim=1, axis=single_axis)
+            return qc_view.squeeze(ndim=1, axis=single_axis)
 
     def __setitem__(self, row_lookup: pandas.Index, col_lookup: pandas.Index, item):
         """
@@ -202,9 +202,9 @@ class _LocationIndexerBase(object):
     def _write_items(self, row_lookup, col_lookup, item):
         """Perform remote write and replace blocks.
         """
-        row_numeric_idx = self.dm.global_idx_to_numeric_idx("row", row_lookup)
-        col_numeric_idx = self.dm.global_idx_to_numeric_idx("col", col_lookup)
-        self.dm.write_items(row_numeric_idx, col_numeric_idx, item)
+        row_numeric_idx = self.qc.global_idx_to_numeric_idx("row", row_lookup)
+        col_numeric_idx = self.qc.global_idx_to_numeric_idx("col", col_lookup)
+        self.qc.write_items(row_numeric_idx, col_numeric_idx, item)
 
 
 class _LocIndexer(_LocationIndexerBase):
@@ -229,13 +229,13 @@ class _LocIndexer(_LocationIndexerBase):
         Returns:
             None
         """
-        if _is_enlargement(row_loc, self.dm.index) or _is_enlargement(
-            col_loc, self.dm.columns
+        if _is_enlargement(row_loc, self.qc.index) or _is_enlargement(
+            col_loc, self.qc.columns
         ):
             _warn_enlargement()
-            self.dm.enlarge_partitions(
-                new_row_labels=self._compute_enlarge_labels(row_loc, self.dm.index),
-                new_col_labels=self._compute_enlarge_labels(col_loc, self.dm.columns),
+            self.qc.enlarge_partitions(
+                new_row_labels=self._compute_enlarge_labels(row_loc, self.qc.index),
+                new_col_labels=self._compute_enlarge_labels(col_loc, self.qc.columns),
             )
 
     def _compute_enlarge_labels(self, locator, base_index):
@@ -277,12 +277,12 @@ class _LocIndexer(_LocationIndexerBase):
     def _compute_lookup(self, row_loc, col_loc) -> Tuple[pandas.Index, pandas.Index]:
         if isinstance(row_loc, list) and len(row_loc) == 1:
             if (
-                isinstance(self.dm.index.values[0], np.datetime64)
+                isinstance(self.qc.index.values[0], np.datetime64)
                 and type(row_loc[0]) != np.datetime64
             ):
                 row_loc = [pandas.to_datetime(row_loc[0])]
-        row_lookup = self.dm.index.to_series().loc[row_loc].index
-        col_lookup = self.dm.columns.to_series().loc[col_loc].index
+        row_lookup = self.qc.index.to_series().loc[row_loc].index
+        col_lookup = self.qc.columns.to_series().loc[col_loc].index
         return row_lookup, col_lookup
 
 
@@ -307,8 +307,8 @@ class _iLocIndexer(_LocationIndexerBase):
         super(_iLocIndexer, self).__setitem__(row_lookup, col_lookup, item)
 
     def _compute_lookup(self, row_loc, col_loc) -> Tuple[pandas.Index, pandas.Index]:
-        row_lookup = self.dm.index.to_series().iloc[row_loc].index
-        col_lookup = self.dm.columns.to_series().iloc[col_loc].index
+        row_lookup = self.qc.index.to_series().iloc[row_loc].index
+        col_lookup = self.qc.columns.to_series().iloc[col_loc].index
         return row_lookup, col_lookup
 
     def _check_dtypes(self, locator):
